requirements.txt
pandas
matplotlib
seaborn
numpy
scikit-learn
transformers
nltk
spacy
torch
torchvision
torchaudio
tensorflow
streamlit
fastapi
uvicorn
jupyter
openpyxl
python-Levenshtein
lightgbm
tf-keras